+++ ubo_app/services/090-web-ui/web-app/src/display/store-event-handler.ts
@@ -11,41 +11,25 @@ import {
   AudioPlayAudioSampleEvent,
   DisplayCompressedRenderEvent,
   Event,
-  Notification,
-  NotificationDisplayType,
-  NotificationsAddAction,
+  DisplayResumeAction,
 } from "../bindings/ubo/v1/ubo_pb";
 
-function showConnectionNotification(store: StoreServiceClient) {
+function requestRedraw(store: StoreServiceClient) {
   const dispatchActionRequest = new DispatchActionRequest();
 
   const action = new Action();
   dispatchActionRequest.setAction(action);
 
-  const notificationsAddAction = new NotificationsAddAction();
-  action.setNotificationsAddAction(notificationsAddAction);
+  const displayResumeAction = new DisplayResumeAction();
+  action.setDisplayResumeAction(displayResumeAction);
 
-  const notification = new Notification();
-  notification.setTitle("Connected to UBO");
-  notification.setContent(
-    "You are now connected to the UBO pod. You can interact with it remotely from this browser session.",
-  );
-  notification.setIcon("󰢹");
-  notification.setDisplayType(
-    NotificationDisplayType.NOTIFICATION_DISPLAY_TYPE_FLASH,
-  );
-  notification.setDismissOnClose(true);
-  notificationsAddAction.setNotification(notification);
-
-  store.dispatchAction(dispatchActionRequest);
+  store.dispatchAction(dispatchActionRequest).then(console.log);
 }
 
 function subscribeToRenderEvents(
   store: StoreServiceClient,
   canvas: HTMLCanvasElement | null,
 ) {
-  showConnectionNotification(store);
-
   const event = new Event();
   event.setDisplayCompressedRenderEvent(new DisplayCompressedRenderEvent());
 
@@ -98,6 +82,8 @@ function subscribeToRenderEvents(
       }
     });
   });
+
+  requestRedraw(store);
 }
 
 export const audioContext = new AudioContext();